Refactor frontend with hooks and components
