+++ frontend/src/App.test.js
@@ -1,8 +1,21 @@
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import AppHeader from './components/AppHeader';
 
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+test('renders header title', () => {
+  Object.defineProperty(window, 'matchMedia', {
+    writable: true,
+    value: jest.fn().mockImplementation(() => ({
+      matches: false,
+      media: '',
+      onchange: null,
+      addListener: jest.fn(),
+      removeListener: jest.fn(),
+      addEventListener: jest.fn(),
+      removeEventListener: jest.fn(),
+      dispatchEvent: jest.fn(),
+    })),
+  });
+  render(<AppHeader totalProducts={0} loading={false} onRefresh={() => {}} />);
+  const headerElement = screen.getByText(/PriceZA Deals Za/i);
+  expect(headerElement).toBeInTheDocument();
 });
